backend/requirements.txt
# Core dependencies for NASA TEMPO Air Quality Pipeline
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

# NASA TEMPO specific dependencies
requests==2.31.0
h5py==3.10.0
s3fs==2023.10.0
pyarrow==14.0.1

# Optional ML dependencies (for advanced bias correction)
# scikit-learn==1.3.2
# scipy==1.11.4

# Background job scheduling (if using schedule library)
# schedule==1.2.0

# For production deployment
gunicorn==21.2.0
numpy==1.24.4
pandas==2.1.4
python-dotenv==1.0.0
email-validator==2.1.0
dnspython==2.4.2
netcdf4==1.6.5
boto3==1.35.0
schedule==1.2.0

# Flask web framework (for APIs)
flask==2.3.3
flask-cors==4.0.0

# Database connectivity
pymysql==1.1.0
mysql-connector-python==8.2.0

# Timezone handling
timezonefinder==6.2.0
pytz==2023.3
